pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sparse-attention-bench"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "matplotlib",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sparse_attention_bench*"]
